Renaming interfaces directories
